Implemented new router api
This new api transforms the router into a type that implements the PSR-15 RequestHandler interface.

Now, state relative to the router process lives in a mutable class called RoutingContext. This is more
efficient in terms of calls.

Lastly, assumptions about containers and other things have been completely removed, because now this router
is strongly typed.